consumer/requirements.txt
kafka-python
minio